Add 'clean' and 'all' for exercise 3 in LCTHW
In Exercise 3 of Learn C The Hard Way,[1] we add to our Makefile such
that `make clean` removes all compiled versions of exercises 1 and 3,
and `make all` compiles the code from exercises 1 and 3.

I ran `make clean` and `make all`, and tweaked minor things about the
programs to lean more.

[1]: http://c.learncodethehardway.org/book/ex3.html